Big refactoring all about printers and 3dprinters.
I did that empty characteristic don't show user. Now massive of
characteristic showed in line. Change looks of admin page of adding new
product. Also add admin opportunity to add 3d printers. Very big
reractoring of controllers. Add ProductDAO.
New design page showed to user.

+++ src/main/java/com/printmaster/nk/service/PrinterService.java
@@ -8,7 +8,7 @@ import com.printmaster.nk.model.SearchPrinters;
 
 public interface PrinterService {
 	 
-    public int addPrinter(Printer p);
+    public long addPrinter(Printer p);
     public void updatePrinter(Printer p);
     public List<Printer> listPrinters();
     public Set<Printer> listSearchPrinters(SearchPrinters searchPrinters);
